Rename package loadfile to chunker.
