+++ dgraph/cmd/bulk/mapper.go
@@ -31,8 +31,8 @@ import (
 	"sync/atomic"
 
 	"github.com/dgraph-io/dgo/protos/api"
+	"github.com/dgraph-io/dgraph/chunker"
 	"github.com/dgraph-io/dgraph/gql"
-	"github.com/dgraph-io/dgraph/loadfile"
 	"github.com/dgraph-io/dgraph/posting"
 	"github.com/dgraph-io/dgraph/protos/pb"
 	"github.com/dgraph-io/dgraph/tok"
